설문지 데이터셋

+++ lib/model/survey_model.dart
@@ -23,7 +23,7 @@ class SurveyResult {
   int? emotion;
   int? awakener;
   List<String>? trackList;
-  List<int>? image;
+  List<double>? image;
   int? afEmotion;
   int? afAwakener;
